Created JdkImmutableCollectionsTest from Collection tests in Jdk8OptionalUsage
