+++ src/test/java/org/mutabilitydetector/jdk/Jdk8OptionalUsage.java
@@ -47,38 +47,6 @@ public class Jdk8OptionalUsage {
 		assertInstancesOf(HasOptionalGeneric.class, areImmutable(), provided("MY_TYPE").isAlsoImmutable());
 	}
 
-	@Test
-	public void usingAListFieldWithListOfIsImmutable() throws Exception {
-		assertInstancesOf(CopyListIntoNewListUsingListOf.class, areImmutable());
-	}
-
-	@Test
-	public void usingASafelyWrappedListFieldWithImmutableElementTypeIsImmutable() throws Exception {
-		assertInstancesOf(HasListOfImmutableThings.class, areImmutable());
-	}
-
-	@Test
-	public void usingAConcreteCollectionFieldWithMutableElementTypeIsMutable() throws Exception {
-		assertInstancesOf(ConcreteCollectionWithMutableElementType.class, areNotImmutable());
-	}
-
-	@Test
-	public void usingAConcreteCollectionFieldWithImmutableElementTypeIsMutable() throws Exception {
-		assertInstancesOf(ConcreteCollectionWithImmutableElementType.class, areImmutable());
-	}
-
-	@Test
-	public void canSuppressWarningsForGenericType() throws Exception {
-		assertInstancesOf(HasImmutableGenericField.class, areNotImmutable());
-		assertInstancesOf(HasImmutableGenericField.class, areImmutable(), provided("T").isAlsoImmutable());
-	}
-
-	@Test
-	public void canSuppressWarningsForCollectionContainingGenericType() throws Exception {
-		assertInstancesOf(HasCollectionFieldContainingGenericElementTypes.class, areNotImmutable());
-		assertInstancesOf(HasCollectionFieldContainingGenericElementTypes.class, areImmutable(), provided("MY_TYPE").isAlsoImmutable());
-	}
-
 	public static final class HasOptionalString {
 		private final Optional<String> optionalStringField;
 
@@ -100,54 +68,6 @@ public class Jdk8OptionalUsage {
 
 		public HasOptionalGeneric(Optional<MY_TYPE> optionalGenericField) {
 			this.optionalGenericField = optionalGenericField;
-		}
-	}
-
-	public static final class ConcreteCollectionWithMutableElementType {
-		private final Set<Date> dates;
-
-		public ConcreteCollectionWithMutableElementType(TreeSet<Date> dates) {
-			this.dates = Collections.unmodifiableSet(new TreeSet(dates));
-		}
-	}
-
-	public static final class ConcreteCollectionWithImmutableElementType {
-		private final Set<String> strings;
-
-		public ConcreteCollectionWithImmutableElementType(TreeSet<String> strings) {
-			this.strings = Collections.unmodifiableSet(new TreeSet(strings));
-		}
-	}
-
-	public static final class CopyListIntoNewListUsingListOf {
-		public final List<String> strings;
-
-		public CopyListIntoNewListUsingListOf(List<String> strings) {
-			this.strings = List.of(strings.toArray(new String[0]));
-		}
-	}
-
-	public static final class HasListOfImmutableThings {
-		public final List<String> strings;
-
-		public HasListOfImmutableThings(List<String> strings) {
-			this.strings = Collections.unmodifiableList(new ArrayList(strings));
-		}
-	}
-
-	public static final class HasImmutableGenericField<T> {
-		public final T t;
-
-		public HasImmutableGenericField(T t) {
-			this.t = t;
-		}
-	}
-
-	public static final class HasCollectionFieldContainingGenericElementTypes<MY_TYPE> {
-		public final List<MY_TYPE> myTypes;
-
-		public HasCollectionFieldContainingGenericElementTypes(List<MY_TYPE> myTypes) {
-			this.myTypes = Collections.unmodifiableList(new ArrayList(myTypes));
 		}
 	}
